# 🚀 improve ctx.QueryParser performance