OLD:
```
Benchmark_Ctx_QueryParser-8       152832              7445 ns/op            1832 B/op         62 allocs/op
Benchmark_Ctx_QueryParser-8       157682              7515 ns/op            1832 B/op         62 allocs/op
Benchmark_Ctx_QueryParser-8       158262              7495 ns/op            1832 B/op         62 allocs/op
Benchmark_Ctx_QueryParser-8       159984              7468 ns/op            1832 B/op         62 allocs/op
```

NEW:
```
Benchmark_Ctx_QueryParser-8       214756              5021 ns/op             912 B/op         39 allocs/op
Benchmark_Ctx_QueryParser-8       232807              5064 ns/op             912 B/op         39 allocs/op
Benchmark_Ctx_QueryParser-8       233775              5111 ns/op             912 B/op         39 allocs/op
Benchmark_Ctx_QueryParser-8       235822              5015 ns/op             912 B/op         39 allocs/op
```